pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "langextract>=1.0.8",
    "openpyxl>=3.1.5",
    "pdfplumber>=0.11.7",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-pptx>=1.0.2",
    "requests>=2.32.5",
]
